bridge_api/requirements.txt
fastapi==0.116.1
httpx==0.28.1
pillow==11.3.0
uvicorn[standard]==0.35.0
